Removed `ContinuousApplicationsController` replacing it with a `ContinuousApplications` concern
Preserves the same functionality
