+++ app/controllers/candidate_interface/course_choices/blocked_submissions_controller.rb
@@ -1,6 +1,7 @@
 module CandidateInterface
   module CourseChoices
-    class BlockedSubmissionsController < ::CandidateInterface::ContinuousApplicationsController
+    class BlockedSubmissionsController < CandidateInterfaceController
+      include CandidateInterface::ContinuousApplications
       def show; end
     end
   end
