+++ app/controllers/candidate_interface/details_controller.rb
@@ -1,5 +1,6 @@
 module CandidateInterface
-  class DetailsController < ContinuousApplicationsController
+  class DetailsController < CandidateInterfaceController
+    include CandidateInterface::ContinuousApplications
     def index
       @application_form_presenter = CandidateInterface::ApplicationFormPresenter.new(current_application)
       @application_cache_key = CacheKey.generate(@application_form_presenter.cache_key_with_version)
